refactor therapy pages
