+++ messages/therapy/en.json
@@ -1,5 +1,5 @@
 {
-  "TherapyBooking": {
+  "Therapy": {
     "title": "Bloom therapy",
     "introduction": "Bumble has partnered with Bloom to offer you free, confidential support. Bloom connects you to a therapist who can offer tailored support. You'll have access to six, hour-long sessions with a therapist of your choice.",
     "bookingButton": "Begin booking",
